refactor: decompose HomeDashboardParts and harden backend error handling
* Extracted components from `HomeDashboardParts.tsx` into individual files in `dashboard/` directory.
* Replaced bare `except Exception:` blocks with specific exceptions (`openai.APIError`, `ValueError`, `OSError`, etc.) in `backend/app/domain/chat/service.py`.
* Verified changes with linter, type-checker, and full test suite across both frontend and backend.
