+++ frontend/app/(main)/home.tsx
@@ -15,7 +15,7 @@ import {
   HomeSectionHeader,
   HomeSurfaceCard,
   HomeTaskRow,
-} from '@/components/home/HomeDashboardParts';
+} from '@/components/home/dashboard';
 import { HomeNewChatModal } from '@/components/home';
 import { HOME_COLORS } from '@/components/home/homeTheme';
 import { formatDate, formatTimeRange, getFirstName, getGreeting, getInitials, isSameDay } from '@/components/home/homeUtils';
