+++ frontend/src/components/home/dashboard/HomeSectionHeader.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import { View } from 'react-native';
+import { AppText } from '@/components/AppText';
+import { ScalePressable } from '@/components/ScalePressable';
+import { HOME_COLORS } from '../homeTheme';
+
+export function HomeSectionHeader({
+  title,
+  actionLabel,
+  onAction,
+}: {
+  title: string;
+  actionLabel?: string;
+  onAction?: () => void;
+}) {
+  return (
+    <View
+      style={{
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginBottom: 12,
+      }}
+    >
+      <AppText variant="h3" style={{ color: HOME_COLORS.text, fontWeight: '700' }}>
+        {title}
+      </AppText>
+      {actionLabel && onAction ? (
+        <ScalePressable onPress={onAction}>
+          <AppText variant="bodySm" style={{ color: HOME_COLORS.primary, fontWeight: '700' }}>
+            {actionLabel}
+          </AppText>
+        </ScalePressable>
+      ) : null}
+    </View>
+  );
+}
