+++ frontend/src/components/home/dashboard/index.ts
@@ -1,0 +1,7 @@
+export * from './HomeHeroCard';
+export * from './HomeAgentBadge';
+export * from './HomeChatRow';
+export * from './HomeTaskRow';
+export * from './HomeActionTile';
+export * from './HomeSurfaceCard';
+export * from './HomeSectionHeader';
