+++ frontend/src/components/home/index.ts
@@ -1,7 +1,6 @@
 export * from './HomeCard';
-export * from './HomeSectionHeader';
 export * from './HomeQuickAction';
 export * from './HomeRecentChatRow';
-export * from './HomeTaskRow';
 export * from './HomeAgentChip';
 export * from './HomeNewChatModal';
+export * from './dashboard';
